+++ src/com/jagex/Class15.java
@@ -1,4 +1,5 @@
 package com.jagex;
+
 final class Class15 implements Interface35 {
 
     public Object method216(byte[] bytes_1, FontMetrics fontmetrics_2, boolean bool_3) {
@@ -35,12 +36,12 @@ final class Class15 implements Interface35 {
         for (int i_4 = 0; i_4 < ints_0.length; i_4++) {
             QuestDefinitions questdefinitions_5 = IndexLoaders.QUEST_LOADER.getQuest(ints_0[i_4]);
             if (questdefinitions_5.graphicId != -1) {
-                NativeSprite nativesprite_6 = (NativeSprite) Class20.aClass229_191.get((long) questdefinitions_5.graphicId);
+                NativeSprite nativesprite_6 = (NativeSprite) Class20.aClass229_191.get(questdefinitions_5.graphicId);
                 if (nativesprite_6 == null) {
                     SpriteDefinitions class91_7 = SpriteDefinitions.getSprite(IndexLoaders.SPRITES_INDEX, questdefinitions_5.graphicId, 0);
                     if (class91_7 != null) {
                         nativesprite_6 = Renderers.SOFTWARE_RENDERER.method8444(class91_7, true);
-                        Class20.aClass229_191.put(nativesprite_6, (long) questdefinitions_5.graphicId);
+                        Class20.aClass229_191.put(nativesprite_6, questdefinitions_5.graphicId);
                     }
                 }
                 if (nativesprite_6 != null) {
